+++ app/Views/hari.php
@@ -65,7 +65,7 @@
                                             data-id-hari="<?= $hari['id_hari'] ?>">
                                         <i data-feather="edit"></i>
                                     </button>
-                                    <a href="/hari/delete/<?= $hari['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"><i data-feather="trash-2"></i></a>
+                                    <a href="/hari/delete/<?= $hari['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"  onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?');"><i data-feather="trash-2"></i></a>
                                 </td>
                             </tr>
                             <?php endforeach; ?>
