+++ app/Views/jam.php
@@ -72,7 +72,7 @@
                                             data-id-jam="<?= $jam['id_jam'] ?>">
                                         <i data-feather="edit"></i>
                                     </button>
-                                    <a href="/jam/delete/<?= $jam['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"><i data-feather="trash-2"></i></a>
+                                    <a href="/jam/delete/<?= $jam['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"  onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?');"><i data-feather="trash-2"></i></a>
                                 </td>
                             </tr>
                             <?php endforeach; ?>
