+++ app/Views/semester.php
@@ -65,7 +65,7 @@
                                             data-id_semester="<?= $semester['id_semester'] ?>" >
                                         <i data-feather="edit"></i>
                                     </button>
-                                    <a href="/semester/delete/<?= $semester['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"><i data-feather="trash-2"></i></a>
+                                    <a href="/semester/delete/<?= $semester['id'] ?>" class="btn btn-datatable btn-icon btn-transparent-dark"  onclick="return confirm('Apakah Anda yakin ingin menghapus data ini?');"><i data-feather="trash-2"></i></a>
                                 </td>
                             </tr>
                             <?php endforeach; ?>
